% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecayRates.R
\name{estimateDecayRates}
\alias{estimateDecayRates}
\title{Estimate Decay Rates}
\usage{
estimateDecayRates(data)
}
\arguments{
\item{data}{The data.frame provided as input to inferNetwork(). Refer to the
data parameter in inferNetwork() for more details.}
}
\value{
Returns a vector of estimated decay rates for each gene.
}
\description{
Code adapted from dynGENIE3.
For each gene, the decay rate is estimated by assuming that the gene expression
x(t) follows: x(t) =  A exp(-alpha * t) + cMin,
between the highest and lowest expression values.
cMin is set to the minimum expression value over all genes and all samples.
}
\examples{
\dontrun{
alphas <- estimateDecayRates(Repressilator)
}

}
\references{
Geurts, P. (2018). dynGENIE3: dynamical GENIE3 for the inference of
gene networks from time series expression data. \emph{Scientific reports},
8(1), 1-12.
}
