% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneCounts}
\alias{GeneCounts}
\title{Gene Counts Dataset}
\format{
A matrix with cols:
\describe{
\item{Cond 1}{Description}
\item{Cond 2}{Description}
}
}
\source{
Univ. Guelph
}
\usage{
GeneCounts
}
\description{
From RNASeq experiment
}
\examples{
\dontrun{
GeneCounts
}
}
\keyword{datasets}
