% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HodgkinHuxley}
\alias{HodgkinHuxley}
\title{Hodgkin-Huxley model of the action potential with no noise.}
\format{
Concentrations are evaluated every 0.1 timesteps for 500 timesteps:
\describe{
\item{V10}{Membrane potential, input current 10}
\item{n10}{Potassium ion channel opening, input current 10}
\item{m10}{Sodium ion channel opening, input current 10}
\item{h10}{Sodium ion channel closing, input current 10}
#' \item{V30}{Membrane potential, input current 30}
\item{n30}{Potassium ion channel opening, input current 30}
\item{m30}{Sodium ion channel opening, input current 30}
\item{h30}{Sodium ion channel closing, input current 30}
#' \item{V50}{Membrane potential, input current 50}
\item{n50}{Potassium ion channel opening, input current 50}
\item{m50}{Sodium ion channel opening, input current 50}
\item{h50}{Sodium ion channel closing, input current 50}
#' \item{V100}{Membrane potential, input current 100}
\item{n100}{Potassium ion channel opening, input current 100}
\item{m100}{Sodium ion channel opening, input current 100}
\item{h100}{Sodium ion channel closing, input current 100}
}
}
\source{
Hodgkin & Huxley (1952)
}
\usage{
HodgkinHuxley
}
\description{
From simulated ODE system. Evaluated at input
currents I = 10, 30, 50, 100 mA/cm^2 evaluated every 0.1 timesteps for 100 timesteps.
}
\examples{
\dontrun{
HodgkinHuxley
}
}
\keyword{datasets}
