% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferSSNetwork.R
\name{inferSSNetwork}
\alias{inferSSNetwork}
\title{Infers a Gene Regulatory Network from Steady-State Data}
\usage{
inferSSNetwork(
  data,
  mask = NULL,
  ntree = 10L,
  mtry = NULL,
  alpha = NULL,
  seed = 777,
  showPlot = FALSE
)
}
\arguments{
\item{data}{A data.frame of gene expression values. Rows are different experiments.
Columns names are gene names.}

\item{mask}{A matrix which only includes the values 1 or NA. Must be of size
numgenes*numgenes. If entry \eqn{(i.j) = 1}, then \eqn{i} can be used in predicting
the value of \eqn{j}. Otherwise, the connection is snipped and such a
dependency is not allowed when training the random forests.}

\item{ntree}{A positive integer indicating the number of trees in each
random forest. Equivalent to the ntree argument in the randomForest package.
Defaults to 10L.}

\item{mtry}{A positive integer indicating the number of randomly sampled
candidates to use at each split of each random forest. Equivalent to the mtry
argument in the randomForest package. Defaults to p/3, where p is the number
of genes. This option is disabled when a mask is provided and the default
value is used.}

\item{alpha}{If not provided, assumed to be 1 for all genes.
If provided, can be a vector of the degradation rates of each gene,
or a single number (same rate for all genes).}

\item{seed}{Random seed for reproducibility. Defaults to 777.}

\item{showPlot}{Plots the weights matrix as a heatmap. Defaults to FALSE.

#' @return Returns an object of class "ugene" with the following items:
\itemize{
  \item network - A matrix storing the importance weights w_ij of each pair
  of genes.
  \item alpha - A vector of the gene product degradation rates,
  possibly inferred from data.
  \item model - A list of "randomForest" objects where model[i] is the
  trained randomForest able to predict changes in concentrations of gene i
  given the current concentrations of all genes.
}}
}
\description{
Given a dataframe genes as columns and different measurements as rows,
returns the adjacency matrix of the inferred network, the estimated
decay rates of each species, and the dynamics of the network learned by \eqn{p}
random forests.
}
\examples{

\dontrun{
   data <- grndata::syntren300.data
   ugene <- inferSSNetwork(data, showPlot = TRUE)
}

}
\references{
Geurts, P. (2018). dynGENIE3: dynamical GENIE3 for the inference of
gene networks from time series expression data. \emph{Scientific reports},
8(1), 1-12.

A. Liaw and M. Wiener (2002). Classification and Regression by
randomForest. R News 2(3), 18--22.
}
