% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrajectory.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Plots individual node dynamics}
\usage{
plotTrajectory(simulation, node.names)
}
\arguments{
\item{simulation}{The result of simulate().}

\item{node.names}{The names of the nodes to plot. Minimum one name.
Maximum six names, as the plot would get cluttered beyond this number.}
}
\description{
Given the result of simulate() (an object of class "simulation"), plots the
dynamics of the specified nodes over time. Specify nodes by their names.
}
\examples{
\dontrun{
  x0 <- Repressilator[1, 2:7]
  ugene <- inferNetwork(Repressilator)
  trajectory <- simulate(ugene, x0)
  plotTrajectory(trajectory, c("p3", "p2", "p1"))
}

}
