% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Repressilator}
\alias{Repressilator}
\title{Repressilator time series with no noise}
\format{
Concentrations are evaluated every 0.1 timesteps for 500 timesteps:
\describe{
\item{t}{Time stamps}
\item{m1}{Concentration of mRNA 1}
\item{p1}{Concentration of protein 1}
\item{m2}{Concentration of mRNA 2}
\item{p2}{Concentration of protein 2}
\item{m3}{Concentration of mRNA 3}
\item{p3}{Concentration of protein 3}
}
}
\source{
Ingalls (2013)
}
\usage{
Repressilator
}
\description{
From simulated ODE system
}
\examples{
\dontrun{
Repressilator
}
}
\keyword{datasets}
