% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundynUGENE.R
\name{rundynUGENE}
\alias{rundynUGENE}
\title{Launch Shiny App For dynUGENE}
\usage{
rundynUGENE()
}
\value{
No return value. Opens the dynUGENE Shiny app.
}
\description{
The shiny app allows users to upload their own data as a .csv file for
network inference, simulation, and tuning.
}
\examples{
\dontrun{
rundynUGENE()
}

}
\references{
Geurts, P. (2018). dynGENIE3: dynamical GENIE3 for the inference of gene
networks from time series expression data. _Scientific reports_, 8(1),
1-12. \href{https://www.nature.com/articles/s41598-018-21715-0}{Link}
}
\author{
Tianyu Lu, \email{tianyu.lu@mail.utoronto.ca}
}
