% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulates an inferred gene regulatory network}
\usage{
simulate(ugene, x0, tend = 100, dt = 0.1, stochastic = FALSE)
}
\arguments{
\item{ugene}{Required. Output of the inferNetwork() function.}

\item{x0}{Required. A data.frame object with a single row, giving the
initial concentrations of all the genes in the network. The order must be the
same as that in the data provided to inferNetwork().}

\item{tend}{Final simulation time. Defaults to 100.}

\item{dt}{Interval between two time steps. Defaults to 0.1. With tend=100,
this implies a total of 1000 time steps, plus the initial concentrations.}

\item{stochastic}{An optional logical argument specifying whether the outputs
of random forests are treated as deterministic (FALSE) or as a distribution
from which a sample is drawn (TRUE). Defaults to FALSE.}
}
\value{
Returns an object of class "simulation" containing the time stamps
in result.t and simulated values of all genes in result.x
}
\description{
Given the output of inferNetwork (object of class "ugene"), simulates the
network given the learned random forests for each node.
}
\examples{
\dontrun{
  x0 <- Repressilator[1, 2:7]
  ugene <- inferNetwork(Repressilator)
  simulate(ugene, x0)
}

}
