% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StochasticHodgkinHuxley}
\alias{StochasticHodgkinHuxley}
\title{Hodgkin-Huxley model of the action potential with diagonal noise}
\format{
Concentrations are evaluated every 0.1 timesteps for 100 timesteps:
\describe{
\item{V}{Membrane potential, input current 10}
\item{n}{Potassium ion channel opening, input current 10}
\item{m}{Sodium ion channel opening, input current 10}
\item{h}{Sodium ion channel closing, input current 10}
}
}
\source{
Hodgkin & Huxley (1952)
}
\usage{
StochasticHodgkinHuxley
}
\description{
From simulated SDE system. Evaluated at
input current I = 10 mA/cm^2 evaluated every 0.1 timesteps for 100 timesteps.
0.05 units of diagonal noise are included for the gating variables n, m, h,
reflecting stochastic ion channel opening and closing.
Five samples from the stochastic process are drawn and are concatenated together
in this dataset. Rows 2-1002 for the first draw, etc.
}
\examples{
\dontrun{
StochasticHodgkinHuxley
}
}
\keyword{datasets}
