% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StochasticRepressilator}
\alias{StochasticRepressilator}
\title{Repressilator time series with diagonal noise.}
\format{
Concentrations are evaluated every 0.1 timesteps for 1000 timesteps:
\describe{
\item{m1}{Concentration of mRNA 1}
\item{p1}{Concentration of protein 1}
\item{m2}{Concentration of mRNA 2}
\item{p2}{Concentration of protein 2}
\item{m3}{Concentration of mRNA 3}
\item{p3}{Concentration of protein 3}
}
}
\source{
Ingalls (2013)
}
\usage{
StochasticRepressilator
}
\description{
From simulated SDE system. Three times the noise is added to mRNA levels than
protein levels. Five
samples from the stochastic process are drawn and are concatenated together
in this dataset. Rows 2-10002 for the first draw, etc.
}
\examples{
\dontrun{
StochasticRepressilator
}
}
\keyword{datasets}
